requirements.txt
pycryptodome>=3.15.0
requests>=2.25.0
